Using repository pattern for match flow.
