+++ data/local-sqldelight/src/commonMain/sqldelight/com/adammcneilly/pocketleague/sqldelight/LocalMatch.sq
@@ -104,3 +104,8 @@ SELECT *
 FROM matchWithEventAndTeams
 WHERE localMatchDateUTC >= DATE()
 ORDER BY localMatchDateUTC ASC;
+
+selectById:
+SELECT *
+FROM matchWithEventAndTeams
+WHERE localMatchId = :matchId;
