+++ shared-test/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/test/TestDataModule.kt
@@ -5,8 +5,7 @@ import com.adammcneilly.pocketleague.data.event.test.FakeEventRepository
 import com.adammcneilly.pocketleague.data.game.GameService
 import com.adammcneilly.pocketleague.data.game.OctaneGGGameService
 import com.adammcneilly.pocketleague.data.local.sqldelight.PocketLeagueDB
-import com.adammcneilly.pocketleague.data.match.MatchService
-import com.adammcneilly.pocketleague.data.match.test.FakeMatchService
+import com.adammcneilly.pocketleague.data.match.MatchRepository
 import com.adammcneilly.pocketleague.data.team.TeamRepository
 import com.adammcneilly.pocketleague.shared.di.DataModule
 
@@ -18,7 +17,8 @@ class TestDataModule : DataModule {
 
     override val eventRepository: EventRepository = FakeEventRepository()
 
-    override val matchService: MatchService = FakeMatchService()
+    override val matchRepository: MatchRepository
+        get() = TODO("Not yet implemented")
 
     override val gameService: GameService = OctaneGGGameService()
